requirements.txt
bleach==1.5.0
bson==0.5.0
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
confluent-kafka==0.11.0
enum34==1.1.6
Flask==0.11.0
Flask-Cors==3.0.3
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.2
Markdown==2.6.10
MarkupSafe==1.0
numpy==1.13.3
protobuf==3.5.1
pyaml==17.10.0
pymongo==3.6.0
pytz==2017.3
PyYAML==3.12
requests==2.18.4
scipy==1.0.0
six==1.11.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
urllib3==1.22
Werkzeug==0.13